src/apps/quizBot/requirements.txt
azure.cosmos
python-telegram-bot
python-dotenv